+++ src/sessionselectiondialog.cpp
@@ -276,7 +276,7 @@ bool SessionSelectionDialog::OnCreate()
     linksPerFileName.clear();
     
     map< boost::posix_time::ptime, wxString, std::greater< boost::posix_time::ptime > > dtToFile;
-    for ( int i = 0 ; i < filesInDir.size() ; ++i )
+    for ( size_t i = 0 ; i < filesInDir.size() ; ++i )
     {
       #ifdef WIN32
       wxString datetime = filesInDir[ i ].AfterLast( '\\' ).AfterFirst( '_' ).Left( 15 );
@@ -319,7 +319,7 @@ bool SessionSelectionDialog::OnCreateNoDialog()
     linksPerFileName.clear();
     
     map< boost::posix_time::ptime, wxString, std::greater< boost::posix_time::ptime > > dtToFile;
-    for ( int i = 0 ; i < filesInDir.size() ; ++i )
+    for ( size_t i = 0 ; i < filesInDir.size() ; ++i )
     {
       #ifdef WIN32
       wxString datetime = filesInDir[ i ].AfterLast( '\\' ).AfterFirst( '_' ).Left( 15 );
